Added a test and readme

+++ Kerem.CodingTracker/Kerem.CodingTracker/Domain/Entities/CodingSession.cs
@@ -5,5 +5,5 @@
         public int Id { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
-        public Decimal Duration { get; set; }
+        public long Duration { get; set; }
     }
